+++ src/components/widgets/PickListPopup/PickListPopup.tsx
@@ -17,6 +17,8 @@ import ColumnTitle from '../../ColumnTitle/ColumnTitle'
 import {RowMetaField} from '../../../interfaces/rowMeta'
 import {buildBcUrl} from '../../../utils/strings'
 import {FieldType} from '../../../interfaces/view'
+import Pagination from '../../ui/Pagination/Pagination'
+import {PaginationMode} from '../../../interfaces/widget'
 
 export interface PickListPopupActions {
     onChange: (payload: ChangeDataItemPayload) => void,
@@ -76,7 +78,19 @@ export const PickListPopup: FunctionComponent<PickListPopupProps & PickListPopup
         [props.pickMap, props.onChange, props.parentBCName, props.cursor]
     )
 
+    const title = React.useMemo(() => <div><h1 className={styles.title}>{props.widget.title}</h1></div>, [props.widget.title])
+    const footer = React.useMemo(() =>
+        <div className={styles.footerContainer}>
+            {!props.widget.options?.hierarchyFull &&
+                <div className={styles.pagination}>
+                    <Pagination bcName={props.widget.bcName} mode={PaginationMode.page} widgetName={props.widget.name}/>
+                </div>
+            }
+        </div>,
+        [props.widget.options?.hierarchyFull, props.widget.bcName, props.widget.name]
+    )
     return <Popup
+        title={title}
         showed={props.showed}
         size="large"
         onOkHandler={props.onClose}
@@ -84,9 +98,9 @@ export const PickListPopup: FunctionComponent<PickListPopupProps & PickListPopup
         bcName={props.widget.bcName}
         widgetName={props.widget.name}
         disablePagination={props.widget.options?.hierarchyFull}
+        footer={footer}
     >
         <div>
-            <h2 className={styles.title}>{props.widget.title}</h2>
             {(props.bcLoading)
             ? <Skeleton loading paragraph={{rows: 5}} />
             : (props.widget.options?.hierarchy || props.widget.options?.hierarchySameBc || props.widget.options?.hierarchyFull)