+++ src/epics/view.ts
@@ -237,8 +237,8 @@ const showAllTableRecordsInit: Epic = (action$, store) => action$.ofType(types.s
 .mergeMap((action) => {
     const resultObservables: Array<Observable<AnyAction>> = []
 
-    const {bcName, route, cursor} = action.payload
-
+    const {bcName, cursor} = action.payload
+    const route = store.getState().router
     resultObservables.push(Observable.of(
         $do.bcChangeCursors({ cursorsMap: { [bcName]: null }})
     ))
@@ -300,7 +300,7 @@ const showAssocPopup: Epic = (action$, store) => action$.ofType(types.showViewPo
                 ...record,
                 id: record.id,
                 _associate: true,
-                _value: record[assocFieldKey]
+                _value: record[action.payload.assocValueKey]
             }
         })
     }